Fix warnings: remove most signed/unsigned comparison warnings
With -Wsign-compare

For the Gtest ASSERT_EQ, opted to add the u suffix to the numeral constant to mark them as unsigned

There's a few remaining which need to be checked or refactored

+++ connection_scan_algorithm/src/alternatives_routing.cpp
@@ -167,7 +167,7 @@ namespace TrRouting
       spdlog::debug("fastest line ids: {}", LinesToString(foundLines));
 
       // Generate combination of group of 1 line, then 2 lines up to the total amount of lines
-      for (int k = 1; k <= foundLines.size(); k++)
+      for (size_t k = 1; k <= foundLines.size(); k++)
       {
         Combinations<std::reference_wrapper<const Line>> combinations(foundLines, k);
 
@@ -180,7 +180,7 @@ namespace TrRouting
       }
 
       // Process all combinations and calculate new route with those excluded
-      for (int i = 0; i < allCombinations.size(); i++)
+      for (size_t i = 0; i < allCombinations.size(); i++)
       {
         if (alternativesCalculatedCount < maxAlternatives && alternativeSequence - 1 < params.maxValidAlternatives)
         {
@@ -222,7 +222,7 @@ namespace TrRouting
                 lastFoundedAtNum = alternativesCalculatedCount;
                 alreadyFoundLines[foundLines] = true;
                 foundLinesTravelTimeSeconds[foundLines] = alternativeCalcResult.totalTravelTime;
-                for (int k = 1; k <= foundLines.size(); k++)
+                for (size_t k = 1; k <= foundLines.size(); k++)
                 {
                   Combinations<std::reference_wrapper<const Line>> combinations(foundLines, k);
                   for (auto newCombination : combinations)
